# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Valentine Hotline API",
    "description": "AI-powered dating screening system API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/avatar/create": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create Avatar",
        "description": "Trigger Tavus avatar creation (placeholder stub for static Heart MVP).",
        "operationId": "create_avatar_api_v1_admin_avatar_create_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarCreateResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/avatar/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Avatar Status",
        "description": "Return Tavus avatar status for the static Heart profile.",
        "operationId": "avatar_status_api_v1_admin_avatar_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TavusStatusInfo"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/calendar/status": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Calendar Status",
        "description": "Return cal.com configuration health and slot preview (placeholder).",
        "operationId": "calendar_status_api_v1_admin_calendar_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarStatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/link/toggle": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Toggle Link",
        "description": "Activate or deactivate the static Heart's public link.",
        "operationId": "toggle_link_api_v1_admin_link_toggle_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkToggleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/dashboard/stats": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Dashboard Stats",
        "description": "Get admin dashboard statistics for the static Heart profile.",
        "operationId": "dashboard_stats_api_v1_admin_dashboard_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/sessions": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Sessions",
        "description": "List all sessions for admin moderation and review.",
        "operationId": "list_sessions_api_v1_admin_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/admin/sessions/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Session Detail",
        "description": "Get full transcript and score detail for one session.",
        "operationId": "session_detail_api_v1_admin_sessions__id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bookings/slots": {
      "get": {
        "tags": [
          "Bookings"
        ],
        "summary": "Get Slots",
        "description": "Get available date booking slots for authenticated suitors (placeholder).",
        "operationId": "get_slots_api_v1_bookings_slots_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailableSlotsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/bookings/create": {
      "post": {
        "tags": [
          "Bookings"
        ],
        "summary": "Create Booking",
        "description": "Create a booking confirmation for an authenticated suitor's session.",
        "operationId": "create_booking_api_v1_bookings_create_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingConfirmationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailableSlotsResponse": {
        "properties": {
          "slots": {
            "items": {
              "$ref": "#/components/schemas/AvailableSlot"
            },
            "type": "array",
            "title": "Slots",
            "description": "Available booking slots."
          }
        },
        "type": "object",
        "title": "AvailableSlotsResponse",
        "description": "List of open booking slots."
      },
      "AvatarCreateResponse": {
        "properties": {
          "replica_id": {
            "type": "string",
            "title": "Replica Id",
            "description": "Replica ID created or reused."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "processing | ready | failed"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message."
          }
        },
        "type": "object",
        "required": [
          "replica_id",
          "status",
          "message"
        ],
        "title": "AvatarCreateResponse",
        "description": "Response for avatar creation trigger endpoint."
      },
      "BookingConfirmationResponse": {
        "properties": {
          "booking_id": {
            "type": "string",
            "format": "uuid",
            "title": "Booking Id",
            "description": "Booking UUID."
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "Related session UUID."
          },
          "heart_id": {
            "type": "string",
            "format": "uuid",
            "title": "Heart Id",
            "description": "Heart UUID."
          },
          "suitor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Suitor Id",
            "description": "Suitor UUID."
          },
          "calcom_booking_id": {
            "type": "string",
            "title": "Calcom Booking Id",
            "description": "External calendar provider booking ID."
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At",
            "description": "Scheduled date-time for the booking."
          },
          "status": {
            "$ref": "#/components/schemas/BookingStatus",
            "description": "Booking lifecycle state."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Booking creation timestamp."
          }
        },
        "type": "object",
        "required": [
          "booking_id",
          "session_id",
          "heart_id",
          "suitor_id",
          "calcom_booking_id",
          "scheduled_at",
          "status",
          "created_at"
        ],
        "title": "BookingConfirmationResponse",
        "description": "Created booking payload."
      },
      "BookingCreateRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "Session UUID being converted to a date booking."
          },
          "slot_datetime": {
            "type": "string",
            "format": "date-time",
            "title": "Slot Datetime",
            "description": "Selected slot datetime in ISO 8601."
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "slot_datetime"
        ],
        "title": "BookingCreateRequest",
        "description": "Create booking payload."
      },
...
      "CalendarStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "connected | not_configured | error"
          },
          "event_type_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Type Id",
            "description": "Configured event type ID."
          },
          "slot_preview": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Slot Preview",
            "description": "Preview list of upcoming slot timestamps."
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CalendarStatusResponse",
        "description": "Calendar status response with lightweight slot preview."
      },
...
      "DashboardStatsResponse": {
        "properties": {
          "total_suitors": {
            "type": "integer",
            "title": "Total Suitors",
            "description": "Total suitors that entered screening."
          },
          "match_rate": {
            "type": "number",
            "title": "Match Rate",
            "description": "Percentage of sessions with a date verdict."
          },
          "avg_scores": {
            "type": "number",
            "title": "Avg Scores",
            "description": "Average weighted score across completed sessions."
          },
          "total_sessions": {
            "type": "integer",
            "title": "Total Sessions",
            "description": "Total sessions created."
          },
          "completed_sessions": {
            "type": "integer",
            "title": "Completed Sessions",
            "description": "Sessions with completed state."
          },
          "upcoming_bookings": {
            "type": "integer",
            "title": "Upcoming Bookings",
            "description": "Count of upcoming date bookings."
          }
        },
        "type": "object",
        "required": [
          "total_suitors",
          "match_rate",
          "avg_scores",
          "total_sessions",
          "completed_sessions",
          "upcoming_bookings"
        ],
        "title": "DashboardStatsResponse",
        "description": "Dashboard summary card metrics."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LinkToggleRequest": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Whether the public heart link should be active."
          }
        },
        "type": "object",
        "required": [
          "is_active"
        ],
        "title": "LinkToggleRequest",
        "description": "Payload to activate/deactivate heart link."
      },
...
      "SessionDetailResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "Session UUID."
          },
          "heart_id": {
            "type": "string",
            "format": "uuid",
            "title": "Heart Id",
            "description": "Heart UUID."
          },
          "suitor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Suitor Id",
            "description": "Suitor UUID."
          },
          "suitor_name": {
            "type": "string",
            "title": "Suitor Name",
            "description": "Suitor display name."
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus",
            "description": "Current session status."
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "Session start timestamp."
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At",
            "description": "Session end timestamp."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Session creation timestamp."
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/SessionDetailTurn"
            },
            "type": "array",
            "title": "Turns",
            "description": "Ordered transcript turns."
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SessionDetailScore"
              },
              {
                "type": "null"
              }
            ],
            "description": "Final score payload when available."
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "heart_id",
          "suitor_id",
          "suitor_name",
          "status",
          "created_at"
        ],
        "title": "SessionDetailResponse",
        "description": "Full session detail response payload."
      },
...
      "SessionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionSummaryItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Session summaries for dashboard list."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total items across all pages."
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number (1-indexed)."
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page",
            "description": "Maximum items per page."
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "per_page"
        ],
        "title": "SessionListResponse",
        "description": "Paginated-style session list response for a heart."
      },
...
      "SessionSummaryItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Session UUID."
          },
          "suitor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Suitor Id",
            "description": "Suitor UUID."
          },
          "suitor_name": {
            "type": "string",
            "title": "Suitor Name",
            "description": "Suitor display name."
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus",
            "description": "Session lifecycle status."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Session creation timestamp."
          },
          "weighted_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weighted Total",
            "description": "Weighted total score if available."
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ],
            "description": "Date/no-date verdict if available."
          }
        },
        "type": "object",
        "required": [
          "id",
          "suitor_id",
          "suitor_name",
          "status",
          "created_at"
        ],
        "title": "SessionSummaryItem",
        "description": "Compact session listing item for dashboard."
      },
...
      "SuccessResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable success message."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SuccessResponse",
        "description": "Generic success message payload."
      },
...
      "TavusStatusInfo": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "not_started | processing | ready | failed"
          },
          "replica_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replica Id",
            "description": "Current Tavus replica ID if available."
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TavusStatusInfo",
        "description": "Tavus replica health information."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Verdict": {
        "type": "string",
        "enum": [
          "date",
          "no_date"
        ],
        "title": "Verdict",
        "description": "Final matchmaking verdict."
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}